+++ src/main/java/chess/controller/ChessViewController.java
@@ -1,0 +1,15 @@
+package chess.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("/chess-game")
+public class ChessViewController {
+
+    @GetMapping()
+    public String init() {
+        return "index";
+    }
+}
