+++ src/main/java/chess/controller/ExceptionAdvice.java
@@ -1,0 +1,16 @@
+package chess.controller;
+
+import chess.dto.ErrorResponseDto;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+@ControllerAdvice
+public class ExceptionAdvice {
+
+    @ExceptionHandler(IllegalArgumentException.class)
+    public ResponseEntity<ErrorResponseDto> handleIllegalArgumentException(Exception exception) {
+        ErrorResponseDto errorResponseDto = new ErrorResponseDto(exception.getMessage());
+        return ResponseEntity.badRequest().body(errorResponseDto);
+    }
+}
